+++ gdpr/src/main/java/com/kairos/response/dto/common/AssessmentBasicResponseDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AssessmentBasicResponseDTO {
 
-    private BigInteger id;
+    private Long id;
     private String name;
     private LocalDate endDate;
     private LocalDate completedDate;
@@ -30,9 +30,9 @@ public class AssessmentBasicResponseDTO {
 
     public void setRisks(List<RiskBasicResponseDTO> risks) { this.risks = risks; }
 
-    public BigInteger getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(BigInteger id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
