+++ gdpr/src/main/java/com/kairos/response/dto/common/AssessmentResponseDTO.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class AssessmentResponseDTO {
 
 
-    private BigInteger id;
+    private Long id;
     private String name;
     private LocalDate endDate;
     private LocalDate completedDate;
@@ -32,9 +32,9 @@ public class AssessmentResponseDTO {
     private AssessmentSchedulingFrequency assessmentSchedulingFrequency;
     private LocalDate startDate;
 
-    public BigInteger getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(BigInteger id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
